backend/cloud_functions/auto_organizer/requirements.txt
google-cloud-bigquery>=3.12.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
numpy>=1.26.0
scikit-learn>=1.4.0
functions-framework>=3.4.0

